followUser, unfollowUser => toggleFollow
