--- src/users/followUser/followUser.typeDefs.js
@@ -1,7 +1,0 @@
-import { gql } from 'apollo-server-express';
-
-export default gql`
-  type Mutation {
-    followUser(username: String!): BaseResponse!
-  }
-`;
