--- src/users/unfollowUser/unfollowUser.typeDefs.js
@@ -1,7 +1,0 @@
-import { gql } from 'apollo-server-express';
-
-export default gql`
-  type Mutation {
-    unfollowUser(username: String!): BaseResponse!
-  }
-`;
